Remove redundant variable names and make them more compact; Introduce multiple TODOs
I made a few simple renames to some variables, mostly with obvious ones like "source_name" to just "name" because it is in the entity "source", and things like "purepecha" to "tsz". Other longer and not trivial names are not changed.
I also added TODOs with things that we need to adress in future commits