[NOT YET TESTED] Factor quality metrics into: processing_status, human_quality and machine rated quality
In this commit I factor the `quality_status` type as a more general `processing_status` which will have a better use for alignment and translation inside the corpus. I separate quality metrics into human based and machine based, so we can note when we are sure that a translation is correct but the model is not sure.